Initial Commit
Currently the program is able to parse the Twitter stream looking for
Keywords or Specific Users. Once a Tweet is found, it will parse it for
a URL and open it in a browser.

Also working is the RSVP OCR system. This takes an RSVP image, finds the
circled hashtag, isolates and reads it with OCR. Currently works 75% of
the time (30/40 tests).