GUI + OCR
Changes have been made to some of the GUI to allow managing of the RSVP
settings as well as code that handles calling the RSVP and Snipe.

I've also added some OCR training things although I may not end up using
them as results are better in some cases but result in loss elsewhere.